Move ComputeDigest to HashHelper
