+++ ProjectLighthouse/Helpers/HashHelper.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
+using System.IO;
 using System.Security.Cryptography;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace LBPUnion.ProjectLighthouse.Helpers {
     [SuppressMessage("ReSharper", "UnusedMember.Global")]
@@ -44,6 +46,34 @@ namespace LBPUnion.ProjectLighthouse.Helpers {
             byte[] bytes = (byte[]) GenerateRandomBytes(256);
 
             return BCryptHash(Sha256Hash(bytes));
+        }
+        
+        public static async Task<string> ComputeDigest(string path, string authCookie, Stream body,
+            string digestKey)
+        {
+            var memoryStream = new MemoryStream();
+
+            var pathBytes = Encoding.UTF8.GetBytes(path);
+            var cookieBytes = string.IsNullOrEmpty(authCookie)
+                ? Array.Empty<byte>()
+                : Encoding.UTF8.GetBytes(authCookie);
+            var keyBytes = Encoding.UTF8.GetBytes(digestKey);
+                
+            await body.CopyToAsync(memoryStream);
+
+            var bodyBytes = memoryStream.ToArray();
+
+            using var sha1 = IncrementalHash.CreateHash(HashAlgorithmName.SHA1);
+            sha1.AppendData(bodyBytes);
+            if (cookieBytes.Length > 0)
+                sha1.AppendData(cookieBytes);
+            sha1.AppendData(pathBytes);
+            sha1.AppendData(keyBytes);
+
+            var digestBytes = sha1.GetHashAndReset();
+            var digestString = Convert.ToHexString(digestBytes).ToLower();
+
+            return digestString;
         }
     }
 }
